Move Dagger.Codegen.Compiler to new file
